# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Pinnacle³ API specification — EB-1A, EB-2 NIW, O-1A immigration consulting platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /criteria:
    get:
      operationId: listCriteria
      tags: [criteria]
      summary: List all USCIS criteria with evidence counts
      parameters:
        - in: query
          name: visaType
          schema:
            type: string
            enum: [eb1a, niw, o1]
          description: Filter by visa type
      responses:
        "200":
          description: List of criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Criterion"
  /criteria/{criterionId}:
    get:
      operationId: getCriterion
      tags: [criteria]
      summary: Get a specific criterion with its evidence
      parameters:
        - in: path
          name: criterionId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Criterion detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CriterionDetail"
        "404":
          description: Criterion not found

  # ━━━ Evidence ━━━
...
  /blueprint:
    get:
      operationId: getBlueprint
      tags: [blueprint]
      summary: Get the user's Elite Blueprint strategy
      responses:
        "200":
          description: Blueprint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Blueprint"
        "404":
          description: No blueprint yet
  /blueprint/milestones:
    get:
      operationId: listMilestones
      tags: [blueprint]
      summary: List blueprint milestones and their completion status
      responses:
        "200":
          description: Milestones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Milestone"

    post:
      operationId: createMilestone
      tags: [blueprint]
      summary: Create a milestone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMilestoneBody"
      responses:
        "201":
          description: Created milestone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Milestone"
  /blueprint/milestones/{milestoneId}:
    patch:
      operationId: updateMilestone
      tags: [blueprint]
      summary: Update a milestone (e.g. mark complete)
      parameters:
        - in: path
          name: milestoneId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMilestoneBody"
      responses:
        "200":
          description: Updated milestone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Milestone"

  # ━━━ Excellence Lab (Courses) ━━━
  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List all Excellence Lab courses
      parameters:
        - in: query
          name: visaType
          schema:
            type: string
            enum: [eb1a, niw, o1]
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
  /courses/{courseId}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get course details and lessons
      parameters:
        - in: path
          name: courseId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseDetail"
        "404":
          description: Course not found
  /courses/{courseId}/progress:
    patch:
      operationId: updateCourseProgress
      tags: [courses]
      summary: Update progress for a course lesson
      parameters:
        - in: path
          name: courseId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCourseProgressBody"
      responses:
        "200":
          description: Updated progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseProgress"

  # ━━━ Dashboard ━━━
...
components:
  schemas:
...
    Criterion:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        visaType:
          type: string
          enum: [eb1a, niw, o1]
        uscisCode:
          type: string
          description: Official USCIS criterion code (e.g. "8 CFR 204.5(h)(3)(i)")
        evidenceCount:
          type: integer
        requiredMinimum:
          type: integer
        strengthLevel:
          type: string
          enum: [none, weak, developing, strong, excellent]
      required:
        - id
        - name
        - description
        - visaType
        - uscisCode
        - evidenceCount
        - requiredMinimum
        - strengthLevel
    CriterionDetail:
      allOf:
        - $ref: "#/components/schemas/Criterion"
        - type: object
          properties:
            evidence:
              type: array
              items:
                $ref: "#/components/schemas/Evidence"
            tips:
              type: string
              description: Guidance for strengthening this criterion
    Evidence:
      type: object
      properties:
        id:
          type: integer
        criterionId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        evidenceType:
          type: string
          enum: [publication, citation, award, salary, membership, media, judging, contribution, letter, other]
        status:
          type: string
          enum: [draft, pending, approved, rejected]
        sourceUrl:
          type: ["string", "null"]
        dateAchieved:
          type: ["string", "null"]
          format: date
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - criterionId
        - title
        - evidenceType
        - status
        - createdAt
        - updatedAt
...
    Blueprint:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        visaTarget:
          type: string
          enum: [eb1a, niw, o1]
        strategySummary:
          type: string
        targetFilingDate:
          type: ["string", "null"]
          format: date
        assignedConsultant:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, active, on_hold, completed]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - visaTarget
        - strategySummary
        - status
        - createdAt
    Milestone:
      type: object
      properties:
        id:
          type: integer
        blueprintId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
          format: date
        completed:
          type: boolean
        completedAt:
          type: ["string", "null"]
          format: date-time
        sortOrder:
          type: integer
      required:
        - id
        - blueprintId
        - title
        - completed
        - sortOrder
    CreateMilestoneBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        dueDate:
          type: string
          format: date
        sortOrder:
          type: integer
      required:
        - title
    UpdateMilestoneBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        dueDate:
          type: string
          format: date
        completed:
          type: boolean
        sortOrder:
          type: integer
    Course:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        visaType:
          type: string
          enum: [eb1a, niw, o1]
        lessonCount:
          type: integer
        estimatedMinutes:
          type: integer
        progressPercent:
          type: integer
        category:
          type: string
      required:
        - id
        - title
        - description
        - visaType
        - lessonCount
        - estimatedMinutes
        - progressPercent
        - category
    CourseDetail:
      allOf:
        - $ref: "#/components/schemas/Course"
        - type: object
          properties:
            lessons:
              type: array
              items:
                $ref: "#/components/schemas/Lesson"
    Lesson:
      type: object
      properties:
        id:
          type: integer
        courseId:
          type: integer
        title:
          type: string
        content:
          type: string
        sortOrder:
          type: integer
        completed:
          type: boolean
        estimatedMinutes:
          type: integer
      required:
        - id
        - courseId
        - title
        - content
        - sortOrder
        - completed
        - estimatedMinutes
    CourseProgress:
      type: object
      properties:
        courseId:
          type: integer
        lessonId:
          type: integer
        completed:
          type: boolean
        progressPercent:
          type: integer
      required:
        - courseId
        - lessonId
        - completed
        - progressPercent
    UpdateCourseProgressBody:
      type: object
      properties:
        lessonId:
          type: integer
        completed:
          type: boolean
      required:
        - lessonId
        - completed
...